Move backend api/controllers/utils tests
